KrishiCare-usingML
Basic Information : This app is based on the fact that the farmer needs his query to be answered and they could be assisted in other ways. This app was developed with respect to the factors such as the resistance to adopt new technology or was unable to read and write or both. This app would pass the queries to the firebase console where there are query handlers who in turn categorize the given queries on the grounds of language, urgency, etc When the query has been categorized, an e-mail would be sent to an expert and he/she would reply to the e-mail as soon as possible. The app is created in three languages which are English Hindi and Marathi (as the target audience was Maharashtrian farmers). A farmer can ask questions in form of text multimedia message and video and image The reply may also be in form of text multimedia message and video and image and any reply will be sent to the farmer s phone. All these stuff are done using Google firebase. Also each activity has an help button which when clicked would open a video tutorial For the ease of the farmer everything will be said in the chosen language of the farmer (till now any out of the three supported languages). The operation of the app is that it can be controlled by voice where by pressing one button it becomes easier for the farmer to work with. This makes it convenient as the phone number of the farmer is used as a login because making email id as an authentication would complicate the process for the farmers. Actually odd buttons are numbered one,two,three and so on and their action is also defined with phrases like this one, that one and so on, thus eliminating language barrier possibility.

PS: There are some components of the app which are not developed completely and these are kept for demonstration purpose.